[MOVE] 모델코드 이동
