+++ PLUB/Sources/Views/Home/Cell/SelectCategory/SelectedCategoryChartCollectionViewCell.swift
@@ -11,67 +11,6 @@ import RxSwift
 import SnapKit
 import Then
 
-struct SelectedCategoryCollectionViewCellModel { // 차트, 그리드일때 둘 다 동일한 모델
-  let plubbingID: Int
-  let name: String
-  let title: String
-  let mainImage: String?
-  let introduce: String
-  var isBookmarked: Bool
-  let selectedCategoryInfoModel: CategoryInfoListModel
-  
-  init(content: Content) {
-    plubbingID = content.plubbingID
-    name = content.name
-    title = content.title
-    mainImage = content.mainImage
-    introduce = content.introduce
-    isBookmarked = content.isBookmarked
-    selectedCategoryInfoModel = .init(
-      placeName: content.placeName,
-      peopleCount: content.remainAccountNum,
-      dateTime: content.days
-      .map { $0.fromENGToKOR() }
-      .joined(separator: ",")
-    + " | "
-    + "(data.time)")
-  }
-  
-  init(content: BookmarkContent) {
-    plubbingID = content.plubbingID
-    name = content.name
-    title = content.title
-    mainImage = content.mainImage
-    introduce = content.introduce
-    isBookmarked = content.isBookmarked
-    selectedCategoryInfoModel = .init(
-      placeName: content.placeName,
-      peopleCount: content.remainAccountNum,
-      dateTime: content.days
-      .map { $0.fromENGToKOR() }
-      .joined(separator: ",")
-    + " | "
-    + "(data.time)")
-  }
-  
-  init(content: SearchContent) {
-    plubbingID = content.plubbingID
-    name = content.name
-    title = content.title
-    mainImage = content.mainImage
-    introduce = content.introduce
-    isBookmarked = content.isBookmarked
-    selectedCategoryInfoModel = .init(
-      placeName: content.placeName,
-      peopleCount: content.remainAccountNum,
-      dateTime: content.days
-      .map { $0.fromENGToKOR() }
-      .joined(separator: ",")
-    + " | "
-    + "(data.time)")
-  }
-}
-
 protocol SelectedCategoryChartCollectionViewCellDelegate: AnyObject {
   func didTappedChartBookmarkButton(plubbingID: Int)
   func updateBookmarkState(isBookmarked: Bool, cell: UICollectionViewCell)
